+++ core/go/internal/publictxmgr/persisted_pub_txs.go
@@ -24,7 +24,7 @@ import (
 
 // public_transactions
 type DBPublicTxn struct {
-	PublicTxnID     uint64                 `gorm:"column:public_txn_id;primaryKey"`
+	PublicTxnID     uint64                 `gorm:"column:pub_txn_id;primaryKey"`
 	From            tktypes.EthAddress     `gorm:"column:from"`
 	Nonce           *uint64                `gorm:"column:nonce"`
 	Created         tktypes.Timestamp      `gorm:"column:created;autoCreateTime:nano"`
@@ -33,11 +33,11 @@ type DBPublicTxn struct {
 	FixedGasPricing tktypes.RawJSON        `gorm:"column:fixed_gas_pricing"`
 	Value           *tktypes.HexUint256    `gorm:"column:value"`
 	Data            tktypes.HexBytes       `gorm:"column:data"`
-	Suspended       bool                   `gorm:"column:suspended"`                                // excluded from processing because it's suspended by user
-	Completed       *DBPublicTxnCompletion `gorm:"foreignKey:signer_nonce;references:signer_nonce"` // excluded from processing because it's done
-	Submissions     []*DBPubTxnSubmission  `gorm:"-"`                                               // we do the aggregation, not GORM
+	Suspended       bool                   `gorm:"column:suspended"`                            // excluded from processing because it's suspended by user
+	Completed       *DBPublicTxnCompletion `gorm:"foreignKey:pub_txn_id;references:pub_txn_id"` // excluded from processing because it's done
+	Submissions     []*DBPubTxnSubmission  `gorm:"-"`                                           // we do the aggregation, not GORM
 	// Binding is used only on queries by transaction (GORM doesn't seem to allow us to define a separate struct for this)
-	Binding *DBPublicTxnBinding `gorm:"foreignKey:signer_nonce;references:signer_nonce;"`
+	Binding *DBPublicTxnBinding `gorm:"foreignKey:pub_txn_id;references:pub_txn_id;"`
 }
 
 func (DBPublicTxn) TableName() string {
@@ -45,7 +45,7 @@ func (DBPublicTxn) TableName() string {
 }
 
 type DBPublicTxnBinding struct {
-	PublicTxnID     uint64                               `gorm:"column:public_txn_id;primaryKey"`
+	PublicTxnID     uint64                               `gorm:"column:pub_txn_id;primaryKey"`
 	Transaction     uuid.UUID                            `gorm:"column:transaction"`
 	TransactionType tktypes.Enum[pldapi.TransactionType] `gorm:"column:tx_type"`
 }
@@ -56,7 +56,7 @@ func (DBPublicTxnBinding) TableName() string {
 
 type DBPubTxnSubmission struct {
 	from            string            `gorm:"-"` // just used to ensure we dispatch to same writer as the associated pubic TX
-	PublicTxnID     uint64            `gorm:"column:public_txn_id;primaryKey"`
+	PublicTxnID     uint64            `gorm:"column:pub_txn_id;primaryKey"`
 	Created         tktypes.Timestamp `gorm:"column:created;autoCreateTime:false"` // we set this as we track the record in memory too
 	TransactionHash tktypes.Bytes32   `gorm:"column:tx_hash"`
 	GasPricing      tktypes.RawJSON   `gorm:"column:gas_pricing"` // no filtering allowed on this field as it's complex JSON gasPrice/maxFeePerGas/maxPriorityFeePerGas calculation
@@ -67,7 +67,7 @@ func (DBPubTxnSubmission) TableName() string {
 }
 
 type DBPublicTxnCompletion struct {
-	PublicTxnID     uint64            `gorm:"column:public_txn_id;primaryKey"`
+	PublicTxnID     uint64            `gorm:"column:pub_txn_id;primaryKey"`
 	Created         tktypes.Timestamp `gorm:"column:created;autoCreateTime:nano"`
 	TransactionHash tktypes.Bytes32   `gorm:"column:tx_hash"`
 	Success         bool              `gorm:"column:success"`
@@ -85,7 +85,7 @@ func (s *DBPubTxnSubmission) WriteKey() string {
 
 type bindingsMatchingSubmission struct {
 	DBPublicTxnBinding `gorm:"embedded"`
-	Submission         *DBPubTxnSubmission `gorm:"foreignKey:signer_nonce;references:signer_nonce;"`
+	Submission         *DBPubTxnSubmission `gorm:"foreignKey:pub_txn_id;references:pub_txn_id;"`
 }
 
 type txFromOnly struct {
